actions bugfix

+++ src/PublicComponentMethods.ts
@@ -4,6 +4,7 @@ export default function getPublicComponentMethods(components, actionsByInstance,
     public name: string;
     public vidoInstance: any;
     public props: any;
+    public destroyed = false;
 
     constructor(instance, vidoInstance, props = {}) {
       this.instance = instance;
@@ -19,14 +20,16 @@ export default function getPublicComponentMethods(components, actionsByInstance,
     /**
      * Destroy component
      */
-    public destroy() {
+    public destroy(): void {
+      if (this.destroyed) return;
       if (this.vidoInstance.debug) {
         console.groupCollapsed(`destroying component ${this.instance}`);
         console.log(clone({ components: components.keys(), actionsByInstance }));
         console.trace();
         console.groupEnd();
       }
-      return this.vidoInstance.destroyComponent(this.instance, this.vidoInstance);
+      this.vidoInstance.destroyComponent(this.instance, this.vidoInstance);
+      this.destroyed = true;
     }
 
     /**
@@ -63,7 +66,7 @@ export default function getPublicComponentMethods(components, actionsByInstance,
      */
     public html(templateProps = {}) {
       const component = components.get(this.instance);
-      if (component) {
+      if (component && !component.destroyed) {
         return component.update(templateProps, this.vidoInstance);
       }
       return undefined;
